+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/Consumption/ConsumptionValidator.cs
