+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/OtherProduction/OtherProductionValidator.cs
@@ -17,9 +17,9 @@ using Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.Rules;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.OtherConsumption
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling.OtherProduction
 {
-    internal class OtherConsumptionMeteringPointValidator : IMasterDataValidatorStrategy
+    internal class OtherProductionValidator : IMasterDataValidatorStrategy
     {
         public BusinessRulesValidationResult CheckRules(MasterData masterData)
         {
